refactor(api): split CSV serialization from worker HTTP helpers
Move rowsToCsv and format negotiation into src/csv-export.mjs so
contracts.mjs can build OpenAPI examples without a circular import
through workers/http.mjs, and extend CSV security tests for coverage.
